+++ tests/e2e/cases/os_user_info/in.ts
@@ -1,6 +1,12 @@
 import * as nodeOs from "node:os";
 import { userInfo } from "os";
 
+let ignoredSeen = "";
+function mark(label: string): string {
+    ignoredSeen += label;
+    return label;
+}
+
 const globalInfo: any = os.userInfo();
 const namespaceInfo: any = nodeOs.userInfo();
 const namedInfo: any = userInfo();
@@ -9,11 +15,12 @@ const UTF8_DASH = "utf-8";
 const GLOBAL_OPTIONS = { encoding: UTF8 } as const;
 const NAMESPACE_OPTIONS = { encoding: UTF8_DASH } as const;
 const NAMED_OPTIONS = { encoding: UTF8 } as const;
-const globalUtf8: any = os.userInfo(GLOBAL_OPTIONS);
-const namespaceUtf8: any = nodeOs.userInfo(NAMESPACE_OPTIONS);
-const namedUtf8: any = userInfo(NAMED_OPTIONS);
+const globalUtf8: any = os.userInfo(GLOBAL_OPTIONS, mark("g"));
+const namespaceUtf8: any = nodeOs.userInfo(NAMESPACE_OPTIONS, mark("n"));
+const namedUtf8: any = userInfo(NAMED_OPTIONS, mark("u"));
 
 console.log("global:", typeof globalInfo.username, typeof globalInfo.homedir, typeof globalInfo.shell, globalInfo.uid >= 0, globalInfo.gid >= 0);
 console.log("namespace:", typeof namespaceInfo.username, namespaceInfo.uid >= 0);
 console.log("named:", typeof namedInfo.username, namedInfo.gid >= 0);
 console.log("options:", typeof globalUtf8.username, typeof namespaceUtf8.homedir, typeof namedUtf8.shell);
+console.log("ignored:", ignoredSeen);